requirements.txt
environs==14.2.0
marshmallow==4.0.0
mysql-connector==2.2.9
paho-mqtt==2.1.0
python-dotenv==1.1.0
